requirements.txt
EasyGoogleTranslate==0.0.4
SeleniumBase==4.25.2
